+++ Kinovea.ScreenManager/Metadata/Drawings/DrawingTimeSegment.cs
@@ -104,7 +104,7 @@ namespace Kinovea.ScreenManager
         #endregion
 
         #region Constructors
-        public DrawingTimeSegment(PointF origin, long timestamp, long averageTimeStampsPerFrame, StyleElements preset = null, IImageToViewportTransformer transformer = null)
+        public DrawingTimeSegment(PointF origin, long timestamp, double averageTimeStampsPerFrame, StyleElements preset = null, IImageToViewportTransformer transformer = null)
         {
             points["a"] = origin;
             points["b"] = origin.Translate(10, 0);
